+++ client/src/main/java/com/influxdb/client/internal/AbstractWriteClient.java
@@ -21,9 +21,7 @@
  */
 package com.influxdb.client.internal;
 
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -67,7 +65,6 @@ import retrofit2.Response;
 public abstract class AbstractWriteClient extends AbstractRestClient implements AutoCloseable {
 
     private static final Logger LOG = Logger.getLogger(AbstractWriteClient.class.getName());
-    private static final List<Integer> ABLE_TO_RETRY_ERRORS = Arrays.asList(429, 503);
     private static final String CLOSED_EXCEPTION = "WriteApi is closed. "
             + "Data should be written before calling InfluxDBClient.close or WriteApi.close.";
     private static final int DEFAULT_WAIT = 30_000;
@@ -264,18 +261,13 @@ public abstract class AbstractWriteClient extends AbstractRestClient implements 
             //
             return source.delay((Function<BatchWriteItem, Flowable<Long>>) pointFlowable -> {
 
-                int delay = jitterDelay();
+                int delay = RetryAttempt.jitterDelay(writeOptions.getJitterInterval());
 
                 LOG.log(Level.FINEST, "Generated Jitter dynamic delay: {0}", delay);
 
                 return Flowable.timer(delay, TimeUnit.MILLISECONDS, scheduler);
             });
         };
-    }
-
-    private int jitterDelay() {
-
-        return (int) (Math.random() * writeOptions.getJitterInterval());
     }
 
     private <T extends AbstractWriteEvent> void publish(@Nonnull final T event) {
@@ -518,49 +510,26 @@ public abstract class AbstractWriteClient extends AbstractRestClient implements 
         Objects.requireNonNull(writeOptions, "WriteOptions are required");
         Objects.requireNonNull(retryScheduler, "RetryScheduler is required");
 
-        return errors -> errors.flatMap(throwable -> {
+        return errors -> errors
+                .zipWith(Flowable.range(1, writeOptions.getMaxRetries() + 1),
+                        (throwable, count) -> new RetryAttempt(throwable, count, writeOptions))
+                .flatMap(attempt -> {
 
-            if (throwable instanceof HttpException) {
+                    Throwable throwable = attempt.getThrowable();
+                    if (attempt.isRetry()) {
 
-                HttpException ie = (HttpException) throwable;
+                        long retryInterval = attempt.getRetryInterval();
 
-                //
-                // The type of error is not able to retry
-                //
-                if (!ABLE_TO_RETRY_ERRORS.contains(ie.code())) {
+                        publish(new WriteRetriableErrorEvent(throwable, retryInterval));
 
+                        return Flowable.just("notify").delay(retryInterval, TimeUnit.MILLISECONDS, retryScheduler);
+                    }
+
+                    //
+                    // This type of throwable is not able to retry
+                    //
                     return Flowable.error(throwable);
-                }
-
-                //
-                // Retry request
-                //
-                long retryInterval;
-
-                String retryAfter = ((HttpException) throwable).response().headers().get("Retry-After");
-                if (retryAfter != null) {
-
-                    retryInterval = TimeUnit.MILLISECONDS.convert(Integer.parseInt(retryAfter), TimeUnit.SECONDS);
-                } else {
-
-                    retryInterval = writeOptions.getRetryInterval();
-
-                    String msg = "The InfluxDB does not specify \"Retry-After\". Use the default retryInterval: {0}";
-                    LOG.log(Level.FINEST, msg, retryInterval);
-                }
-
-                retryInterval = retryInterval + jitterDelay();
-
-                publish(new WriteRetriableErrorEvent(throwable, retryInterval));
-
-                return Flowable.just("notify").delay(retryInterval, TimeUnit.MILLISECONDS, retryScheduler);
-            }
-
-            //
-            // This type of throwable is not able to retry
-            //
-            return Flowable.error(throwable);
-        });
+                });
     }
 
     @Nonnull